Microcontroller: Basic Port Access
I'm at the Recurse Center after Shenzhen I/O came out, so I have
the time and the resources to make a Python simulator and
compiler for the game.

Design files are in plaintext, which is what you'd do if you
wanted people to write a compiler for your game...

Maybe I should attempt TIS-100 first, but the machine language
in TIS-100 is completely different, so I wouldn't be able to
reuse any of the code.