Microcontroller: CPU Redesign; not instruction.
CPU is no longer a Singleton.

Microcontrollers have CPUs.

Compiles to a list instead of a tree.

Test instructions shouldn't be recursive.  Now we can keep track
of a linear cursor, which  makes jumps a lot easier.

Also it turns out that test instructions just switch all
conditional instructions on the chip on or off, not just those
chained immediately following it.  The previous implementation was
incorrect.

Threw in a not instruction just for fun.

Thanks to @AndreaOrru for pairing. #NeverGraduate