Microcontroller: jmp instruction.
Now the execute command keeps track of the current cursor.

If a cursor is returned from an instruction, the cursor will be
changed.

Otherwise, it'll just be incremented.

Since we don't want to put the system into infinite loops, exec
ends after running out of instructions.

We'll add a generator to do the actual step through.