Microcontroller: Step; I/O Test Code.
This isn't actually the behavior we need.

In the real world, the test instructions are running at the exact
same time, so the write should return 0.

So how to handle time is still a bit of an issue.  A new thread for
each CPU, maybe, but we don't actually need these to run in
realtime, and even then, manually stepping each Microprocessor
should still yield the parallel result.

We'll probably need to have some outer context that also keeps
track of time, and then maybe the registers also have to store
write and read timestamps internally...

I'm still mulling it over.