Board to run processors with an understanding of time.
Turns out time is super hard to implement.

Still want to run more tests on this, and I still have to make it
so that a write on a GPIO and a read to a connected port in the
same cycle doesn't return the new value until the next cycle.