Board: Write has no effect on parallel read.
Since both the actions are happening at the same time, there's no
way for the read instruction to get the value immediately.  It
effectively hasn't happened yet.

Also added a Board instance to Microcontroller in order to get
around Microcontrollers not on a Board getting confused, since
straight access and code execution doesn't advance any cycles
outside of the Board context.

In this case, the lack of test isolation is a good thing; it keeps
us from getting into a strange state with the global time stuff.

And of course this is why global stuff is terrible in general.

I'm thinking of refactoring so that ports don't have a link
function and Circuits live on the Board and not the
Microcontroller.

I can't think of an instance where it even makes sense to have
Microcontrollers and Circuits outside of the Board context,
except for testing.  But then again, this entire library is
really just a bunch of unit tests right now.